requirements.txt
absl-py==2.3.1
accelerate==1.12.0
astunparse==1.6.3
atomicwrites==1.4.1
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
deepface==0.0.96
filelock==3.20.0
fire==0.7.1
Flask==3.1.2
flask-cors==6.0.1
flatbuffers==25.9.23
fonttools==4.61.0
fsspec==2025.10.0
gast==0.6.0
gdown==5.2.0
google-pasta==0.2.0
grpcio==1.76.0
gunicorn==23.0.0
h5py==3.15.1
huggingface-hub==0.36.0
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
joblib==1.5.2
keras==3.12.0
kiwisolver==1.4.9
libclang==18.1.1
lightecc==0.0.3
lightphe==0.0.17
lz4==4.4.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.4
mpmath==1.3.0
mtcnn==1.0.0
namex==0.1.0
networkx==3.6
numpy==2.2.6
opencv-contrib-python==4.11.0.86
opencv-python==4.12.0.88
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
polars==1.35.2
polars-runtime-32==1.35.2
protobuf==6.33.1
psutil==7.1.3
py==1.11.0
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.5
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
retina-face==0.0.17
rich==14.2.0
safetensors==0.7.0
scipy==1.16.3
sentencepiece==0.2.1
setuptools==80.9.0
six==1.17.0
sounddevice==0.5.3
soupsieve==2.8
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
tf_keras==2.20.1
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.3
typing_extensions==4.15.0
tzdata==2025.2
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==2.0.1
